(PC-16911)[PRO] chore: delete TableFrame component
