--- pro/src/screens/Bookings/BookingsRecapTable/components/Table/TableFrame/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TableFrame'
