+++ pro/src/screens/Bookings/BookingsRecapTable/components/index.ts
@@ -2,4 +2,4 @@ export * from './CellsFormatter'
 export * from './Filters'
 export { default as NoFilteredBookings } from './NoFilteredBookings'
 export { default as Header } from './Header'
-export { default as TableFrame } from './Table/TableFrame'
+export { default as TableWrapper } from './Table/TableWrapper'
